Test value of pid before issuing kill() command. This rarely resulting in slurmd killing itself if the KILL_JOB RPC arrived before the job began execution (the pid in the data structure was still zero.

+++ trunk/src/slurmd/no_interconnect.c
@@ -66,8 +66,12 @@ int interconnect_postfini (slurmd_job_t *job)
 	/*
 	 *  Kill all processes in the job's session
 	 */
-	debug2("Sending SIGKILL to pgid %d", job->smgr_pid); 
-	kill(-job->smgr_pid, SIGKILL);
+	if (job->smgr_pid) {
+		debug2("Sending SIGKILL to pgid %d", job->smgr_pid); 
+		kill(-job->smgr_pid, SIGKILL);
+	} else
+		debug("Job %u.%u: Bad pid valud %ld", job->jobid, 
+		      job->stepid, job->smgr_pid);
 
 	return SLURM_SUCCESS;
 }
